requirements.txt
discord~=1.0.1
requests~=2.32.2
python-dotenv~=0.14.0
datetime~=5.0
regex==2023.10.3